Use new test syntax for file app composables
